+++ src/carbon.cpp
@@ -1,5 +1,5 @@
 #include <Rcpp.h>
-#include "hydraulics.h"
+#include "tissuemoisture.h"
 #include "biophysicsutils.h"
 using namespace Rcpp;