+++ src/hydraulics.h
@@ -25,8 +25,6 @@ NumericVector rootxylemConductanceProportions(NumericVector V, NumericVector L);
 
 NumericVector psi2Weibull(double psi50, double psi88 = NA_REAL, double psi12 = NA_REAL);
 
-double stemWaterCapacity(double Al2As, double height, double wd);
-double leafWaterCapacity(double SLA, double ld);
 
 NumericVector regulatedPsiTwoElements(double Emax, double psiSoil, double krhizomax, double kxylemmax, double n, double alpha, double c, double d, double dE = 0.1, double psiMax = -10.0);