+++ src/tissuemoisture.cpp
@@ -7,6 +7,40 @@
 using namespace Rcpp;
 
 
+/**
+ * Calculate water capacity of stem per leaf area (in mm = l·m-2)
+ * 
+ * Al2As - Leaf area to sapwood area ratio (in m2·m-2)
+ * height - plant height (in cm)
+ * wd - wood density (in g·cm-3)
+ * http://www.fao.org/forestry/17109/en/
+ */
+// [[Rcpp::export("moisture_sapwoodWaterCapacity")]]
+double sapwoodWaterCapacity(double Al2As, double height, NumericVector V, NumericVector L, double wd) {
+  int nlayers = V.size();
+  double woodPorosity = (1.0- (wd/1.54));
+  double vAbove = 1000*(height/(Al2As*100.0))*woodPorosity;
+  double vBelow = 0.0;
+  for(int i=0;i<nlayers;i++) {
+    vBelow += 1000*(V[i]*(L[i]/10.0)/(Al2As*100.0))*woodPorosity;
+  }
+  return(vAbove+vBelow);
+  // 
+  // return(1000*(height/(Al2As*100.0))*(1.0- (wd/1.54)));
+}
+
+/**
+ * Calculate  capacity of leaves per leaf area (in mm = l·m-2)
+ * 
+ * SLA - Specific leaf area (in m2/kg)
+ * ld - leaf density (in g/cm3 = 1000 kg/m3)
+ */
+// [[Rcpp::export("moisture_leafWaterCapacity")]]
+double leafWaterCapacity(double SLA, double ld) {
+  return(1000.0/(1000.0*ld*SLA))*(1.0- (ld/1.54));
+}
+
+
 // [[Rcpp::export("moisture_turgorLossPoint")]]
 double turgorLossPoint(double pi0, double epsilon) {
   return((pi0*epsilon)/(pi0+epsilon));
@@ -86,8 +120,6 @@ double tissueRelativeWaterContent(double psiSym, double pi0, double epsilon,
   double apo_rwc = std::min(1.0-femb, apoplasticRelativeWaterContent(psiApo, c, d)); //Water content cannot be higher than the fraction of non-embolized conduits
   return(sym_rwc*(1.0-af)+apo_rwc*af);
 }
-
-
 
 /**
  * Translates soil water balance results to fuel moisture content of plant cohorts.