+++ NEWS.md
@@ -5,6 +5,7 @@
 # Version 2.7.0
 - Basic water balance model with relative water content
 - 'pRootDisc' eliminated from species parameters
+- Plant water balance and cuticular tranpiration added to the basic water balance model
 
 # Version 2.6.2
 - Control parameter 'modifyInput' is no longer available. Functions 'spwb', 'pwb' and 'growth' do not modify input objects and return an element 'spwbOutput' or 'growthOutput' with a copy of the final state.